refactor(server): extract graceful shutdown to boot/boot-shutdown.ts (#4243 step 4)
Extract the ~200-line graceful shutdown handler from server.ts main()
into boot/boot-shutdown.ts with registerShutdownHandler().

The handler tears down all services in reverse dependency order:
1. Drain + broadcast shutdown SSE frame
2. Close HTTP server (with grace period)
3. Stop monitors, intervals, config watcher
4. Destroy JSONL watcher, pipelines, memory bridge
5. Kill all CC sessions
6. Stop managed services via ServiceContainer
7. Save metrics + metering + analytics cache
8. Flush audit log + OpenTelemetry spans
9. Cleanup PID file

Changes:
- New: src/boot/boot-shutdown.ts (282 lines)
- Modified: src/server.ts (1196 → 969, -227 lines, ~19% reduction)
- Removed 5 unused imports from server.ts
- Mutable shutdownLateRefs pattern for pidFilePath (set after listenWithRetry)

Stacks on #4322 (ACP boot) and #4323 (metering boot).

Verification:
- tsc --noEmit: clean
- 5652 tests passed (390 test files)
- No behavioral changes — pure extraction

Part of #4243
